feat : PlayPopup 일부 작업 완료
1. 화면 터치 시 소지금액 증가
2. Text, Button 바인드 및 기본 함수 연동
3. 탭 클릭시 탭 변경 완료
-> Lotto 탭 Content가 벗어나는 문제 해결 요망

현재 WorkTab, LottoTab
임시 데이터 생성 후 연동 중

+++ I Will be a Rich/Assets/Scripts/Data/StatData.cs
@@ -20,7 +20,7 @@ public class StatData
     [XmlAttribute]
     public int value;
     [XmlAttribute]
-    public int increaseValue;
+    public float increaseValue;
     [XmlAttribute]
     public int maxLevel;
 }
